Jenkinsfile, new S3 endpoint withPathStyleAccess
We are not using amazon, so we need to set our own endpoint. Would like
to allow the endpoint to be set in config, but my quick attempt didn't
work. Seeing as are endpoint is set in stone, I don't think it's worth
the effort at the moment. Will have to rebuild for a enw endpoint though

Also had to change the path style for buckets. Also not a configurable
change. It's hardcoded for now.

Added Jenkins file so we can build automatically. Not sure if it works
yet, have to see if this commit builds :)